Include \R macro in output
The test for "special" macros in reconstruct() gave a false-positive
for "\\R" when "\\R" should be treated like "\\cr". This caused
reconstruct() to omit the "\\R" in the roxygen comments.

Add test for NULL tags. Without this test `tag(rd) != "\\R"` and
`tag(rd) == toupper(tag(rd))` are both logical(0). The result of
`logical(0) && logical(0)` is NA, and ifelse() returns NA if the test
argument is NA. So both the prefix and suffix would contain NA.

Fixes #34.